Add redux and middlewares to the project
