+++ app/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux"
+import events from "./eventReducer"
+import user from "./userReducer"
+
+export default rootReducer = combineReducers({
+	events,
+	user
+})
