+++ package.json
@@ -7,11 +7,18 @@
     "test": "jest"
   },
   "dependencies": {
+    "axios": "^0.15.3",
+    "nock": "^9.0.2",
     "react": "15.4.1",
     "react-native": "0.38.1",
     "react-native-drawer": "^2.3.0",
     "react-native-router-flux": "^3.37.0",
-    "react-native-scrollable-tab-view": "^0.7.1"
+    "react-native-scrollable-tab-view": "^0.7.1",
+    "react-redux": "^5.0.2",
+    "redux": "^3.6.0",
+    "redux-logger": "^2.7.4",
+    "redux-mock-store": "^1.2.1",
+    "redux-thunk": "^2.2.0"
   },
   "jest": {
     "preset": "react-native"
